FlexTreeの名前などをわかりやすくリファクタ
