+++ src/spatial_id/collection/flex_tree/map/tests/snapshots/kasane_logic__spatial_id__collection__flex_tree__map__tests__golden__golden_tests__persisted_byte_length_is_stable.snap.new
@@ -1,0 +1,6 @@
+---
+source: src/spatial_id/collection/flex_tree/map/tests/golden.rs
+assertion_line: 83
+expression: bytes.len()
+---
+400
