+++ src/spatial_id/collection/flex_tree/map/tests/snapshots/kasane_logic__spatial_id__collection__flex_tree__map__tests__golden__golden_tests__persisted_bytes_are_stable.snap.new
@@ -1,0 +1,30 @@
+---
+source: src/spatial_id/collection/flex_tree/map/tests/golden.rs
+assertion_line: 76
+expression: hex_dump(&bytes)
+---
+0000  01 00 00 00 00 00 00 00 00 00 01 01 00 00 00 00 
+0010  00 00 00 00 00 0b 01 00 00 00 00 00 00 00 00 0a 
+0020  02 00 00 00 00 00 00 00 01 01 00 00 00 00 00 00 
+0030  00 00 00 0b 04 00 00 00 00 00 00 00 00 0a 00 00 
+0040  00 00 05 00 00 00 00 09 03 00 00 00 06 00 00 00 
+0050  00 08 07 00 00 00 00 00 00 00 00 07 08 00 00 00 
+0060  00 00 00 00 01 02 00 00 00 00 00 00 00 00 00 0b 
+0070  00 00 00 00 0a 00 00 00 00 0a 0b 00 00 00 00 00 
+0080  00 00 00 09 0c 00 00 00 00 00 00 00 00 08 0d 00 
+0090  00 00 00 00 00 00 01 03 00 00 00 00 00 00 00 00 
+00a0  00 09 00 00 00 00 0f 00 00 00 00 08 00 00 00 00 
+00b0  10 00 00 00 00 07 0e 00 00 00 11 00 00 00 00 06 
+00c0  09 00 00 00 12 00 00 00 00 05 13 00 00 00 00 00 
+00d0  00 00 00 04 14 00 00 00 00 00 00 00 01 03 00 00 
+00e0  00 00 00 00 00 00 00 09 16 00 00 00 00 00 00 00 
+00f0  00 08 00 00 00 00 17 00 00 00 00 07 00 00 00 00 
+0100  18 00 00 00 00 06 19 00 00 00 00 00 00 00 00 05 
+0110  1a 00 00 00 00 00 00 00 00 04 1b 00 00 00 00 00 
+0120  00 00 00 03 15 00 00 00 1c 00 00 00 00 02 1d 00 
+0130  00 00 00 00 00 00 00 01 1e 00 00 00 00 00 00 00 
+0140  00 00 1f 00 00 00 00 00 00 00 aa bb cc ee fc ff 
+0150  ff ff 01 00 00 00 f5 ff ff ff 02 00 00 00 ef ff 
+0160  ff ff 01 00 00 00 01 00 00 00 00 00 20 00 00 00 
+0170  01 00 00 00 00 00 02 00 00 00 00 02 00 00 00 00 
+0180  80 fe ff ff 21 00 00 00 c6 ff ff ff 03 00 00 00
